fix: `kong restart` starts kong even if not previously running

+++ src/cli/restart.lua
@@ -10,11 +10,10 @@ Options:
   -c,--config (default %s) configuration file
 ]], constants.CLI.GLOBAL_KONG_CONF))
 
--- Check if running, will exit if not
-signal.is_running(args.config)
-
-if not signal.send_signal(args.config, "stop") then
-  cutils.logger:error_exit("Could not stop Kong")
+if signal.is_running(args.config) then
+  if not signal.send_signal(args.config, "stop") then
+    cutils.logger:error_exit("Could not stop Kong")
+  end
 end
 
 signal.prepare_kong(args.config)
